code status as of end of second class
