+++ README.md
@@ -5,7 +5,7 @@ A repo for the Foundations II: JavaScript class at Code Fellows.
 Need a name for your band, or just a random band name to liven up your day? This is the site for you.
 
 To play with this repo:
-1.) Fork it
-2.) Clone it to your machine
-3.) ```cd``` into the band-name-generator folder
-4.) Run ```npm install```
+- Fork it
+- Clone it to your machine
+- ```cd``` into the band-name-generator folder
+- Run ```npm install```
