+++ app/index.html
@@ -17,14 +17,10 @@
     <h3>Submit words!</h3>
     <form id="submitWords">
       <label>Give us an adjective:</label><input type="text" name="adjective">
-      <label>Give us a verb:</label><input type="text" name="verb">
-      <label>Give us a plural noun:</label><input type="text" name="noun">
       <input type="submit" value="Submit Words"/>
     </form>
 
     <p id='adjectiveRes'></p>
-    <p id='verbRes'></p>
-    <p id='nounRes'></p>
 
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
     <script src="script.js"></script>
