--- lib/verb.js
@@ -1,16 +1,0 @@
-'use strict';
-
-var Verb = function() {
-  this.sleeping = true;
-  this.skiing = true;
-  this.flopped = true;
-  this.waiting = true;
-  this.plotting = true;
-  this.burping = true;
-  this.sneezing = true;
-  this.flying = true;
-  this.snoozing = true;
-  this.dancing = true;
-};
-
-module.exports = Verb;
