+++ server.js
@@ -5,14 +5,9 @@ var app = express();
 var bodyparser = require('body-parser');
 var port = process.env.PORT || 3000;
 var Adjective = require('./lib/adjective.js');
-var Verb = require('./lib/verb.js');
-var Noun = require('./lib/noun.js');
 var getRandomWord = require('./lib/getRandomWord.js');
-var postWord = require('./lib/postWord.js');
 
 var adjective = new Adjective();
-var verb = new Verb();
-var noun = new Noun();
 
 app.use(bodyparser.json());
 app.use(bodyparser.urlencoded({
@@ -20,6 +15,44 @@ app.use(bodyparser.urlencoded({
 }));
 
 app.use(express.static(__dirname + '/app/'));
+
+var Verb = function() {
+  this.sleeping = true;
+  this.skiing = true;
+  this.flopped = true;
+  this.waiting = true;
+  this.plotting = true;
+  this.burping = true;
+  this.sneezing = true;
+  this.flying = true;
+  this.snoozing = true;
+  this.dancing = true;
+};
+var verb = new Verb();
+
+var Noun = function() {
+  this.ninjas = true;
+  this.cowboys = true;
+  this.bears = true;
+  this.beats = true;
+  this.alligators = true;
+  this.hamsters = true;
+  this.otters = true;
+  this.helicopters = true;
+  this.unicorns = true;
+  this.rhymes = true;
+};
+var noun = new Noun();
+
+function postWord (word, wordObject) {
+  if (wordObject.hasOwnProperty(word)) {
+    return {msg: 'We already have your awesome word, ' + word + ', in our list.'};
+  }
+
+  wordObject[word] = true;
+  console.dir(wordObject);
+  return {msg: 'Thanks for submitting ' + word + '!'};
+};
 
 app.get('/adjective', function(req, res) {
   res.json(getRandomWord(adjective));
@@ -35,16 +68,6 @@ app.get('/noun', function(req, res) {
 
 app.post('/adjective', function(req, res) {
   var word = postWord(req.body.word, adjective);
-  res.json(word);
-});
-
-app.post('/verb', function(req, res) {
-  var word = postWord(req.body.word, verb);
-  res.json(word);
-});
-
-app.post('/noun', function(req, res) {
-  var word = postWord(req.body.word, noun);
   res.json(word);
 });
